# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Error Database API
  description: |
    A comprehensive API for the Error Database application. This API provides endpoints for managing error codes, solutions, user authentication, and advanced features like bookmarks, notifications, achievements, webhooks, and data export.

    ## Features
    - **Error Management**: Create, read, update, and search error codes
    - **Solution Management**: Submit, vote on, and verify solutions
    - **User Management**: Authentication, profiles, and achievements
    - **Bookmarks**: Save solutions for later reference
    - **Notifications**: Real-time in-app notifications
    - **Webhooks**: External integrations with event subscriptions
    - **Analytics**: Search analytics and usage statistics
    - **Export**: Data export in JSON and CSV formats

    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
  version: 1.0.0
  contact:
    name: Error Database Team
    email: support@errordb.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3010/api
    description: Local development server
  - url: https://api.errordb.com/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string

  # Search
...
  /export/stats:
    get:
      tags:
        - Export
      summary: Get export statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Export statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalErrors:
                        type: integer
                      totalSolutions:
                        type: integer
                      totalUsers:
                        type: integer
                      totalApplications:
                        type: integer
                      totalCategories:
                        type: integer
                      dateRange:
                        type: object
                        properties:
                          earliest:
                            type: string
                            format: date-time
                          latest:
                            type: string
                            format: date-time
  /export/errors:
    get:
      tags:
        - Export
      summary: Export errors (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Exported data
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
        '403':
          description: Admin access required
  /export/solutions:
    get:
      tags:
        - Export
      summary: Export solutions (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Exported data
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
  /export/user-data:
    get:
      tags:
        - Export
      summary: Export user data
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data export
          content:
            application/json:
              schema:
                type: object
  /export/analytics:
    get:
      tags:
        - Export
      summary: Export analytics (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Analytics export
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
            refreshToken:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
        reputation:
          type: integer
        solutionCount:
          type: integer
        verifiedSolutionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string